requirements.txt
pydub
openai
python-dotenv